gemini integrated in frontend and fixed UI
